+++ distant-core/src/manager/data/destination.rs
@@ -1,18 +1,11 @@
 use crate::serde_str::{deserialize_from_str, serialize_to_str};
-use derive_more::{Display, Error, From};
 use serde::{de::Deserializer, ser::Serializer, Deserialize, Serialize};
-use std::{convert::TryFrom, fmt, hash::Hash, str::FromStr};
-use uriparse::{
-    Authority, AuthorityError, Host, Password, Scheme, URIReference, URIReferenceError, Username,
-    URI,
-};
+use std::{fmt, hash::Hash, str::FromStr};
 
-/// Represents an error that occurs when trying to parse a destination from a str
-#[derive(Copy, Clone, Debug, Display, Error, From, PartialEq, Eq)]
-pub enum DestinationError {
-    MissingHost,
-    URIReferenceError(URIReferenceError),
-}
+mod host;
+mod parser;
+
+pub use host::{Host, HostParseError};
 
 /// `distant` connects and logs into the specified destination, which may be specified as either
 /// `hostname:port` where an attempt to connect to a **distant** server will be made, or a URI of
@@ -20,93 +13,31 @@ pub enum DestinationError {
 ///
 /// * `distant://hostname:port` - connect to a distant server
 /// * `ssh://[user@]hostname[:port]` - connect to an SSH server
+///
+/// **Note:** Due to the limitations of a URI, an IPv6 address is not supported.
 #[derive(Clone, Debug, Hash, PartialEq, Eq)]
-pub struct Destination(URIReference<'static>);
+pub struct Destination {
+    /// Sequence of characters beginning with a letter and followed by any combination of letters,
+    /// digits, plus (+), period (.), or hyphen (-) representing a scheme associated with a
+    /// destination
+    pub scheme: Option<String>,
+
+    /// Sequence of alphanumeric characters representing a username tied to a destination
+    pub username: Option<String>,
+
+    /// Sequence of alphanumeric characters representing a password tied to a destination
+    pub password: Option<String>,
+
+    /// Consisting of either a registered name (including but not limited to a hostname) or an IP
+    /// address. IPv4 addresses must be in dot-decimal notation, and IPv6 addresses must be
+    /// enclosed in brackets ([])
+    pub host: Host,
+
+    /// Port tied to a destination
+    pub port: Option<u16>,
+}
 
 impl Destination {
-    /// Returns a reference to the scheme associated with the destination, if it has one
-    pub fn scheme(&self) -> Option<&str> {
-        self.0.scheme().map(Scheme::as_str)
-    }
-
-    /// Replaces the current scheme of the destination with the provided scheme, returning the old
-    /// scheme as a string if it existed
-    pub fn replace_scheme(&mut self, scheme: &str) -> Result<Option<String>, URIReferenceError> {
-        self.0
-            .set_scheme(Some(Scheme::try_from(scheme).map(Scheme::into_owned)?))
-            .map(|s| s.map(|s| s.to_string()))
-    }
-
-    /// Returns the host of the destination as a string
-    pub fn to_host_string(&self) -> String {
-        // NOTE: We guarantee that there is a host for a destination during construction
-        self.0.host().unwrap().to_string()
-    }
-
-    /// Returns the port tied to the destination, if it has one
-    pub fn port(&self) -> Option<u16> {
-        self.0.port()
-    }
-
-    /// Returns the username tied with the destination if it has one
-    pub fn username(&self) -> Option<&str> {
-        self.0.username().map(Username::as_str)
-    }
-
-    /// Returns the password tied with the destination if it has one
-    pub fn password(&self) -> Option<&str> {
-        self.0.password().map(Password::as_str)
-    }
-
-    /// Replaces the host of the destination
-    pub fn replace_host(&mut self, host: &str) -> Result<(), URIReferenceError> {
-        let username = self
-            .0
-            .username()
-            .map(Username::as_borrowed)
-            .map(Username::into_owned);
-        let password = self
-            .0
-            .password()
-            .map(Password::as_borrowed)
-            .map(Password::into_owned);
-        let port = self.0.port();
-        let _ = self.0.set_authority(Some(
-            Authority::from_parts(
-                username,
-                password,
-                Host::try_from(host)
-                    .map(Host::into_owned)
-                    .map_err(AuthorityError::from)
-                    .map_err(URIReferenceError::from)?,
-                port,
-            )
-            .map(Authority::into_owned)
-            .map_err(URIReferenceError::from)?,
-        ))?;
-        Ok(())
-    }
-
-    /// Indicates whether the host destination is globally routable
-    pub fn is_host_global(&self) -> bool {
-        match self.0.host() {
-            Some(Host::IPv4Address(x)) => {
-                !(x.is_broadcast()
-                    || x.is_documentation()
-                    || x.is_link_local()
-                    || x.is_loopback()
-                    || x.is_private()
-                    || x.is_unspecified())
-            }
-            Some(Host::IPv6Address(x)) => {
-                // NOTE: 14 is the global flag
-                x.is_multicast() && (x.segments()[0] & 0x000f == 14)
-            }
-            Some(Host::RegisteredName(name)) => !name.trim().is_empty(),
-            None => false,
-        }
-    }
-
     /// Returns true if destination represents a distant server
     pub fn is_distant(&self) -> bool {
         self.scheme_eq("distant")
@@ -118,15 +49,10 @@ impl Destination {
     }
 
     fn scheme_eq(&self, s: &str) -> bool {
-        match self.0.scheme() {
-            Some(scheme) => scheme.as_str().eq_ignore_ascii_case(s),
+        match self.scheme.as_ref() {
+            Some(scheme) => scheme.eq_ignore_ascii_case(s),
             None => false,
         }
-    }
-
-    /// Returns reference to inner [`URIReference`]
-    pub fn as_uri_ref(&self) -> &URIReference<'static> {
-        &self.0
     }
 }
 
@@ -136,72 +62,62 @@ impl AsRef<Destination> for &Destination {
     }
 }
 
-impl AsRef<URIReference<'static>> for Destination {
-    fn as_ref(&self) -> &URIReference<'static> {
-        self.as_uri_ref()
+impl AsMut<Destination> for &mut Destination {
+    fn as_mut(&mut self) -> &mut Destination {
+        *self
     }
 }
 
 impl fmt::Display for Destination {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
+        if let Some(scheme) = self.scheme.as_ref() {
+            write!(f, "{scheme}://")?;
+        }
+
+        if let Some(username) = self.username.as_ref() {
+            write!(f, "{username}")?;
+        }
+
+        if let Some(password) = self.password.as_ref() {
+            write!(f, ":{password}")?;
+        }
+
+        if self.username.is_some() || self.password.is_some() {
+            write!(f, "@")?;
+        }
+
+        write!(f, "{}", self.host)?;
+
+        if let Some(port) = self.port {
+            write!(f, ":{port}")?;
+        }
+
+        Ok(())
     }
 }
 
 impl FromStr for Destination {
-    type Err = DestinationError;
+    type Err = &'static str;
 
+    /// Parses a destination in the form `[scheme://][[username][:password]@]host[:port]`
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // Disallow empty (whitespace-only) input as that passes our
-        // parsing for a URI reference (relative with no scheme or anything)
-        if s.trim().is_empty() {
-            return Err(DestinationError::MissingHost);
-        }
-
-        let mut destination = URIReference::try_from(s)
-            .map(URIReference::into_owned)
-            .map(Destination)
-            .map_err(DestinationError::URIReferenceError)?;
-
-        // Only support relative reference if it is a path reference as
-        // we convert that to a relative reference with a host
-        if destination.0.is_relative_reference() {
-            let path = destination.0.path().to_string();
-            destination.replace_host(path.as_str())?;
-            let _ = destination.0.set_path("/")?;
-        }
-
-        Ok(destination)
-    }
-}
-
-impl<'a> TryFrom<URIReference<'a>> for Destination {
-    type Error = DestinationError;
-
-    fn try_from(uri_ref: URIReference<'a>) -> Result<Self, Self::Error> {
-        if uri_ref.host().is_none() {
-            return Err(DestinationError::MissingHost);
-        }
-
-        Ok(Self(uri_ref.into_owned()))
-    }
-}
-
-impl<'a> TryFrom<URI<'a>> for Destination {
-    type Error = DestinationError;
-
-    fn try_from(uri: URI<'a>) -> Result<Self, Self::Error> {
-        let uri_ref: URIReference<'a> = uri.into();
-        Self::try_from(uri_ref)
+        parser::parse(s)
     }
 }
 
 impl FromStr for Box<Destination> {
-    type Err = DestinationError;
+    type Err = &'static str;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let destination = s.parse::<Destination>()?;
         Ok(Box::new(destination))
+    }
+}
+
+impl<'a> PartialEq<&'a str> for Destination {
+    #[allow(clippy::cmp_owned)]
+    fn eq(&self, other: &&'a str) -> bool {
+        self.to_string() == *other
     }
 }
 
@@ -228,47 +144,14 @@ mod tests {
     use super::*;
 
     #[test]
-    fn parse_should_fail_if_string_is_only_whitespace() {
-        let err = "".parse::<Destination>().unwrap_err();
-        assert_eq!(err, DestinationError::MissingHost);
-
-        let err = " ".parse::<Destination>().unwrap_err();
-        assert_eq!(err, DestinationError::MissingHost);
-
-        let err = "\t".parse::<Destination>().unwrap_err();
-        assert_eq!(err, DestinationError::MissingHost);
-
-        let err = "\n".parse::<Destination>().unwrap_err();
-        assert_eq!(err, DestinationError::MissingHost);
-
-        let err = "\r".parse::<Destination>().unwrap_err();
-        assert_eq!(err, DestinationError::MissingHost);
-
-        let err = "\r\n".parse::<Destination>().unwrap_err();
-        assert_eq!(err, DestinationError::MissingHost);
-    }
-
-    #[test]
-    fn parse_should_succeed_with_valid_uri() {
-        let destination = "distant://localhost".parse::<Destination>().unwrap();
-        assert_eq!(destination.scheme().unwrap(), "distant");
-        assert_eq!(destination.to_host_string(), "localhost");
-        assert_eq!(destination.as_uri_ref().path().to_string(), "/");
-    }
-
-    #[test]
-    fn parse_should_fail_if_relative_reference_that_is_not_valid_host() {
-        let _ = "/".parse::<Destination>().unwrap_err();
-        let _ = "/localhost".parse::<Destination>().unwrap_err();
-        let _ = "my/path".parse::<Destination>().unwrap_err();
-        let _ = "/my/path".parse::<Destination>().unwrap_err();
-        let _ = "//localhost".parse::<Destination>().unwrap_err();
-    }
-
-    #[test]
-    fn parse_should_succeed_with_nonempty_relative_reference_by_setting_host_to_path() {
-        let destination = "localhost".parse::<Destination>().unwrap();
-        assert_eq!(destination.to_host_string(), "localhost");
-        assert_eq!(destination.as_uri_ref().path().to_string(), "/");
+    fn display_should_output_using_available_components() {
+        let destination = Destination {
+            scheme: None,
+            username: None,
+            password: None,
+            host: Host::Name("example.com".to_string()),
+            port: None,
+        };
+        assert_eq!(destination, "example.com");
     }
 }